feature: 基于 SPI 自定义序列化器

+++ example-common/src/main/java/com/wth/example/common/service/UserService.java
@@ -13,7 +13,7 @@ public interface UserService {
      */
     User getUser(User user);
 
-    default short getNumber() {
+    default long getNumber() {
         return 1;
     }
